+++ pkg/deployment/deployment_inspector.go
@@ -121,6 +121,12 @@ func (d *Deployment) inspectDeployment(lastInterval util.Interval) util.Interval
 			d.CreateEvent(k8sutil.NewErrorEvent("Member failure detection failed", err, d.apiObject))
 		}
 
+		// Immediate actions
+		if err := d.reconciler.CheckDeployment(); err != nil {
+			hasError = true
+			d.CreateEvent(k8sutil.NewErrorEvent("Reconciler immediate actions failed", err, d.apiObject))
+		}
+
 		// Create scale/update plan
 		if err := d.reconciler.CreatePlan(); err != nil {
 			hasError = true
@@ -163,6 +169,12 @@ func (d *Deployment) inspectDeployment(lastInterval util.Interval) util.Interval
 		if err := d.createAccessPackages(); err != nil {
 			hasError = true
 			d.CreateEvent(k8sutil.NewErrorEvent("AccessPackage creation failed", err, d.apiObject))
+		}
+
+		// Ensure deployment bootstrap
+		if err := d.EnsureBootstrap(); err != nil {
+			hasError = true
+			d.CreateEvent(k8sutil.NewErrorEvent("Bootstrap failed", err, d.apiObject))
 		}
 
 		// Inspect deployment for obsolete members