+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.3.10](---) (XXXX-XX-XX)
 - Added Pod Disruption Budgets for all server groups in production mode.
+- Added Priority Class Name to be specified per server group.
 
 ## [0.3.9](https://github.com/arangodb/kube-arangodb/tree/0.3.9) (2019-02-28)
 [Full Changelog](https://github.com/arangodb/kube-arangodb/compare/0.3.8...0.3.9)